feat: add stringRequest method
stringRequest lets you utilize the class options, error handling and other benefits to using gqlr, but you are in charge of creating the string body sent to the graphql endpoint.  This is helpful when working with libraries like SWR, where you need to stringify js objects into a cache key, that is then passed to your fetcher function.